Project26_InvoiceGen/requirements.txt
fpdf2==2.7.8
